Further updates to the PCode->Crucible bridge. We can now translate all the code for the simple "fib.xml" example.
The entire PCode program is translated into a single crucible function CFG; this
makes possible a uniform translation of jump/call and branchind/return, which
are semantically identical according to the PCode documentation.  The generated crucible
function takes arguments giving the initial state of the register file and RAM.

In many ways, the trickiest part of this translation is the need to support
indirect jumps, which crucible does not currently support.  For now, I am making
the "whole program" assumption; that is, I assume all relevant indirect jump targets
are included in the PCode disassembly.  This allows me to construct a trampoline
rather than needing real indirect jumps.  The trampoline examines the address of
the desired indirect jump and, one-by-one, compares it to all the disassembled
instructions present in the PCode program, branching to the correct one when it is found.
If the program jumps to a location unknown to the trampoline, we return from the
crucible function, under the assumption that an unknown indirect jump indicates a return
to the calling context.

Still to be done:
  * Arrange for the RAM image to contain the bytes specified in the PCode data segment
  * Return the modified machine state when we fall off the bottom of the trampoline
  * Additional unit tests for the Java crucible API
  * Run some translated programs on concrete inputs for a basic sanity check
  * Run some translated programs on symbolic inputs!
  * Enhance the Crucible->SAWCore backend to handle the necessary additional primitives